+++ src/main/java/com/tikie/shiro/mapper/RoleMapper.java
@@ -5,14 +5,14 @@ import com.tikie.test.mapper.MyBatisRepository;
 import org.springframework.data.repository.query.Param;
 
 /**
- * @targget     RoleMapper
+ *              RoleMapper
  *
  * @author      tikie
- * @date        2016-10-09
+ *              2016-10-09
  * @version     1.0.0
  */
 @MyBatisRepository
 public interface RoleMapper {
 
-    Role getById(@Param("id") Integer id);
+    Role getById(@Param("id") Long id);
 }
