+++ src/main/java/com/tikie/shiro/service/RoleService.java
@@ -3,13 +3,13 @@ package com.tikie.shiro.service;
 import com.tikie.shiro.entity.Role;
 
 /**
- * @targget     RoleMapper
+ *              RoleMapper
  *
  * @author      tikie
- * @date        2016-10-09
+ *              2016-10-09
  * @version     1.0.0
  */
 public interface RoleService {
 
-    Role getById(Integer id);
+    Role getById(Long id);
 }
